lot more work on maps/sets + some refactoring and benchmarking...
